Remove GrSecurity/PAX and trim down SELinux
Now that GrSecurity/PaX are no longer available for free download, it is
almost impossible to get it without paying significant amounts of money.
Remove them from the recommendation, but mention that it remains a
viable option for anyone who has a subscription.

Additionally, trim down the SELinux section to remove the detailed
instructions on audit2allow and staff_u. Such details are probably
best suited for a dedicated document.

Signed-off-by: Konstantin Ryabitsev <konstantin@linuxfoundation.org>